% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  app_mode = c("regular", "rt"),
  app_title = "Crowd Forecast",
  data,
  first_forecast_date = "auto",
  horizons = 4,
  horizon_interval = 7,
  submission_date = NA,
  selection_vars = c("location", "target_type"),
  forecast_quantiles = c(0.01, 0.025, seq(0.05, 0.95, 0.05), 0.975, 0.99),
  google_account_mail = NULL,
  path_service_account_json = NULL,
  forecast_sheet_id,
  user_data_sheet_id,
  user_management = TRUE,
  past_forecasts = NULL,
  app_up_to_date = TRUE,
  default_distribution = "log-normal",
  default_baseline = "constant baseline",
  force_increasing_uncertainty = TRUE,
  ...
)
}
\arguments{
\item{forecast_quantiles}{the quantiles for which a forecast shall be made}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
