% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_from_rt.R
\name{simulate_cases_from_rt}
\alias{simulate_cases_from_rt}
\title{Simulate cases from Rt forecast}
\usage{
simulate_cases_from_rt(forecast, epinow2_fit, num_samples = 100)
}
\arguments{
\item{forecasts}{the crowdforecastr forecast reactive list}
}
\value{
A data frame
}
\description{
Simulate cases from Rt forecast
}
