% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  data,
  horizons = 4,
  submission_date = NA,
  selection_vars = c("location", "target_type"),
  forecast_quantiles = c(0.01, 0.025, seq(0.05, 0.95, 0.05), 0.975, 0.99),
  google_account_mail = NULL,
  forecast_sheet_id,
  user_data_sheet_id,
  user_management = TRUE,
  path_past_forecasts = NULL,
  app_up_to_date = TRUE,
  ...
)
}
\arguments{
\item{forecast_quantiles}{the quantiles for which a forecast shall be made}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
